frontend : corrections

+++ frontend/src/app/components/account-detail/account-detail.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import {ActivatedRoute, Router} from "@angular/router";
 import {UserService} from "../../services/user.service";
 import {User} from "../../models/user.model";
+import {MatDialog} from "@angular/material/dialog";
+import {ConfirmDialogComponent} from "../confirm-dialog/confirm-dialog.component";
 
 @Component({
   selector: 'app-account-detail',
@@ -16,7 +18,8 @@ export class AccountDetailComponent implements OnInit {
 
   constructor(private activated: ActivatedRoute,
               private userService: UserService,
-              private router: Router) { }
+              private router: Router,
+              public dialog: MatDialog) { }
 
   ngOnInit(): void {
 
@@ -44,14 +47,22 @@ export class AccountDetailComponent implements OnInit {
     this.router.navigate(['account'])
   }
 
-  toDelete() {
-    this.userService.deleteUser(+this.userId)
-      .subscribe(() => {
-        this.router.navigate(['home'])
-      })
+  onConfirmDialog() {
+    const dialogRef = this.dialog.open(ConfirmDialogComponent)
+    dialogRef.afterClosed().subscribe(formData => {
+      if (!formData) {
+        return 'Suppression annulée !'
+      }
+      this.userService.deleteUser(+this.userId)
+        .subscribe(() => {
+          this.router.navigate(['home'])
+        })
+      return 'Suppression confirmée !'
+    })
   }
 
   toHome() {
     this.router.navigate(['home'])
   }
+
 }
